Move credit for inspiration from README to KDoc comment
The specific rendering techniques currently in use are not that
central to the project anymore, as I plan to experiment with
different rendering techniques. Also, the technique has been used
in the gaming industry much earlier than I thought, being known as
"Mode 7" and originally invented for early gaming consoles.

Since I still want to give credit to Notch for his live stream and
programming skills, I refer to the source of inspiration closer to
where it is actually used, in the KDoc comment of the
`ViewportRenderer` class.