Prevent use of java.lang.Math via ArchUnit test
The goal of this project is to reinvent the wheel in many cases, like
creating our own math library and implementing the math we need when
we need it.

Currently this test fails, since we still use the standard library
in some cases, but at this point in the project we can simply focus
on making this architecture test pass again ASAP by replacing these
method calls with custom implementations.