Remove architecture rule to restrict math lib usage
We can easily recreate some of the math library functionality, but
some of the more complicated functions are not worth the effort. So
we might recreate some of the math functions for fun and practice,
but it's ok to use the standard library math for, e.g., trigonometric
functions etc.