Render triangle bounding boxes with nice flickering effect
We are not yet determining whether each point lies within the
triangle, but generating a random boolean instead, which leads to a
nice flickering effect (which we of course don't want to keep except
temporarily).
We are however already minimizing the rendering to each triangle's
bounding box, which currently (also temporarily) results in an
interesting pixelating effect.