Inline use of tickTime
Since we now also have tickTime in addition to elapsedTime, we can
now also call it from the comparator used for sorting. This was not
possible with elapsedTime, since that would have broken the contract
of the comparator/selector used for sorting due to yielding different
values across different comparator calls. This made it necessary to
save the elapsedTime in a local variable and pass it to each compara-
tor call.