Animate rotation in vertex shader
This not only separates the concerns better and allows for using
arbitrary vertex shaders, but it is also more performant since we
now calculate the rotation only once. Previously we had to rotate
the light source and the (global) z values for z buffering in
addition to rotating the mesh's vertices.