Fix unintended recursion in property initializer
The initializer called the Vec3f contructor, leading to an uninteded
recursion and stack overflow. Now we use a getter instead, which
works since the getter is not called upon object initialization, but
only on demand. Lazy initialization would work as well, but is
probably not beneficial and might have an overhead that is not worth
it with such a trivial calculation.