Remove coroutine for full screen and window closing
The coroutine was used to automatically switch to full screen and
back to windowed, and then close the window a bit later.
This was helpful during implementation of this functionality, but is
not needed anymore now. Game developers can use this functionality to
programmatically change the screen settings or close the game from
within game logic (e.g., a game's pause/main menu).