Make Swing window non-resizable
We will control the window size via the render resolution and not the other
way around. If we want to make it possible to have resizable windows on some
platforms, we can add this later, but we will then need to handle auto-resolution
(based on window size) vs. explicitly configured resolution (as is common in most
games).