Implement bilinear texture filtering
For now we do this directly in the fragment shader. Later on, we can
move this into a separate class (e.g., TextureSampler, maybe some of
the logic for color interpolation directly into the color).

The choice is between "nearest" and "bilinear" filtering. "Nearest"
rounds the UV coordinate to the nearest texel in the texture. This is
very fast. "Bilinear" bilinearly interpolates between the 4 pixels
surrounding the current UV coordinates, using the distance from each
of those 4 pixels as the interpolation weights.

We are not handling sub-pixel precision, and we are using the pixels'
bottom-left corner rather than their center for simplicity.