Fix calculation of TBN matrix to fix lighting issue
The w component for the 3 vectors was 1, which is wrong in this case,
because the TBN matrix is normally a 3x3 matrix and we are using a
4x4 matrix that behaves like a 3x3 matrix by setting the w components
to 0 and only the m44 position to 1.

We also normalize the vectors just in case, because the 3D model data
is not guaranteed to provide normalized vectors, but this was not part
of the lighting problem in this case.