Keep sun static and floor unlit for performance
Interpolation seems to be the performance bottleneck, making the floor
very performance intensive despite its simple geometry, since it covers
so many pixels on screen. We set it to unlit for now, and add a shader
that only interpolates UVs so that we can use prebaked lighting for
large surfaces like floors and walls.