Greatly improve framerate
We implement a performance hack: Only rasterize only one in every
four pixels, then denoise the image in the end in post-processing.
We furthermore disable gamma correction. It is still supported by
the engine and can be added by a game, but it is a trade-off that
is likely not worth it for most games in a software renderer,
expect maybe in tech demos that focus on visual quality over
performance.